Finished tests.
Notes:
Test Order Array - it could be more efficient by casting to Number
whilst sorting, but I challenge the result changing the type of the
elements and would like to discuss this. (hence adding the traversal
cast rather than integrating it into the function)

Test Get Diff Array - Set implementation would go elsewhere; it was
simple enough to patch in here rather than go npm exploring

Test Get Distance - I think the tested result may be wrong; I attempted
3 different approaches and never got to the same number, and indeed
didn't get it on websites online.
Furthermore, the type was wrong and was testing a string when the
distance would/should be a Number or Float.
Regardless, since I didn't know the use case, I chose to implement the
most "accurate" (i.e.  closest to the value provided in the test file)
and use chai-almost to express the expectation of accuracy, in this case
to tens of meters.

Cheers.